Moved GUI into one package
