+++ src/main/java/GUI/StartFromBeginning/GUI_View.java
@@ -1,9 +1,5 @@
 package GUI.StartFromBeginning;
 
-import GUI.StartFromBeginning.Done.ActionTabsPanel;
-import GUI.StartFromBeginning.Done.QuickPopupActionPanel;
-import GUI.StartFromBeginning.Done.StartUpPanel;
-
 import javax.swing.*;
 import java.awt.*;
